modular Vite config
